// src/shared/types/records.types.ts
export type IndebtedPeople = {
  _id?: string;
  name: string;
  amount: number;
  amountPaid: number;
  isPaid: boolean;
}

export type IndebtedPeopleUI = IndebtedPeople & {
  amountFormatted: string
  amountPaidFormatted: string;
  remainingAmountFormatted: string;
}

export type AccountRecord = {
  _id: string;
  transferRecord?: TransferRecord;
  userId: string;
  shortName: string;
  typeOfRecord: TypeOfRecord;
  description: string;
  amount: number;
  amountFormatted: string;
  date: Date;
  fullDate: string;
  formattedTime: string;
  category: Category | null;
  subCategory: string;
  tag: string[];
  indebtedPeople: IndebtedPeople[];
  account: string;
  budgets: string[];
}

export type ExpenseRecord = AccountRecord & {
  isPaid: boolean;
}

export type IncomeRecord = AccountRecord & {
  expensesPaid: ExpensePaid[];
}

export type ExpensePaid = {
  _id: string;
  shortName: string;
  amount: number;
  amountFormatted: string;
  formattedTime: string;
  fullDate: string;
  isPaid: boolean;
  date: Date;
}

export type DeleteRecordPayload = {
  recordId: string;
}
// ...
export type CreateExpensePayload = {
  account: string;
  amount: number;
  budgets: string[];
  category: string;
  date: Date;
  description: string;
  indebtedPeople: IndebtedPeople[];
  isPaid: boolean;
  linkedBudgets: string[];
  shortName: string;
  subCategory: string;
  tag: string[];
  typeOfRecord: 'expense'
}
export type EditExpensePayload = CreateExpensePayload & {
  recordId: string;
}
// ...
export interface ExpenseDataResponse {
  data: {
    expense: ExpenseRecord
  }
  error: null;
  message: string[];
  success: boolean;
  version: string;
}